Refactor SolanaIndexer to listen to changes to artist_coins
