+++ config/solana_config.go
@@ -19,6 +19,7 @@ type SolanaConfig struct {
 	GrpcToken    string
 
 	MintAudio solana.PublicKey
+	MintUSDC  solana.PublicKey
 
 	RewardManagerProgramID   solana.PublicKey
 	RewardManagerState       solana.PublicKey
@@ -35,6 +36,7 @@ const (
 	// Dev
 	DevSolanaRelay                   = "http://audius-protocol-discovery-provider-1/solana/relay"
 	DevMintAudio                     = "37RCjhgV1qGV2Q54EHFScdxZ22ydRMdKMtVgod47fDP3"
+	DevMintUSDC                      = "26Q7gP8UfkDzi7GMFEQxTJaNJ8D2ybCUjex58M5MLu8y"
 	DevRewardManagerProgramID        = "testLsJKtyABc9UXJF8JWFKf1YH4LmqCWBC42c6akPb"
 	DevRewardManagerState            = "DJPzVothq58SmkpRb1ATn5ddN2Rpv1j2TcGvM3XsHf1c"
 	DevRewardManagerLookupTable      = "GNHKVSmHvoRBt1JJCxz7RSMfzDQGDGhGEjmhHyxb3K5J"
@@ -45,6 +47,7 @@ const (
 	// Stage
 	StageSolanaRelay                   = "https://discoveryprovider.staging.audius.co/solana/relay"
 	StageMintAudio                     = "BELGiMZQ34SDE6x2FUaML2UHDAgBLS64xvhXjX5tBBZo"
+	StageMintUSDC                      = "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZy4z6cQ"
 	StageRewardManagerProgramID        = "CDpzvz7DfgbF95jSSCHLX3ERkugyfgn9Fw8ypNZ1hfXp"
 	StageRewardManagerState            = "GaiG9LDYHfZGqeNaoGRzFEnLiwUT7WiC6sA6FDJX9ZPq"
 	StageRewardManagerLookupTable      = "ChFCWjeFxM6SRySTfT46zXn2K7m89TJsft4HWzEtkB4J"
@@ -55,6 +58,7 @@ const (
 	// Prod
 	ProdSolanaRelay                   = "https://discoveryprovider.audius.co/solana/relay"
 	ProdMintAudio                     = "9LzCMqDgTKYz9Drzqnpgee3SGa89up3a247ypMj2xrqM"
+	ProdMintUSDC                      = "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZy4z6cQ"
 	ProdRewardManagerProgramID        = "DDZDcYdQFEMwcu2Mwo75yGFjJ1mUQyyXLWzhZLEVFcei"
 	ProdRewardManagerState            = "71hWFVYokLaN1PNYzTAWi13EfJ7Xt9VbSWUKsXUT8mxE"
 	ProdRewardManagerLookupTable      = "4UQwpGupH66RgQrWRqmPM9Two6VJEE68VZ7GeqZ3mvVv"
@@ -91,6 +95,7 @@ func NewSolanaConfig() SolanaConfig {
 	case "":
 		cfg.SolanaRelay = DevSolanaRelay
 		cfg.MintAudio = solana.MustPublicKeyFromBase58(DevMintAudio)
+		cfg.MintUSDC = solana.MustPublicKeyFromBase58(DevMintUSDC)
 		cfg.RewardManagerProgramID = solana.MustPublicKeyFromBase58(DevRewardManagerProgramID)
 		cfg.RewardManagerState = solana.MustPublicKeyFromBase58(DevRewardManagerState)
 		cfg.RewardManagerLookupTable = solana.MustPublicKeyFromBase58(DevRewardManagerLookupTable)
@@ -102,6 +107,7 @@ func NewSolanaConfig() SolanaConfig {
 	case "staging":
 		cfg.SolanaRelay = StageSolanaRelay
 		cfg.MintAudio = solana.MustPublicKeyFromBase58(StageMintAudio)
+		cfg.MintUSDC = solana.MustPublicKeyFromBase58(StageMintUSDC)
 		cfg.RewardManagerProgramID = solana.MustPublicKeyFromBase58(StageRewardManagerProgramID)
 		cfg.RewardManagerState = solana.MustPublicKeyFromBase58(StageRewardManagerState)
 		cfg.RewardManagerLookupTable = solana.MustPublicKeyFromBase58(StageRewardManagerLookupTable)
@@ -113,6 +119,7 @@ func NewSolanaConfig() SolanaConfig {
 	case "production":
 		cfg.SolanaRelay = ProdSolanaRelay
 		cfg.MintAudio = solana.MustPublicKeyFromBase58(ProdMintAudio)
+		cfg.MintUSDC = solana.MustPublicKeyFromBase58(ProdMintUSDC)
 		cfg.RewardManagerProgramID = solana.MustPublicKeyFromBase58(ProdRewardManagerProgramID)
 		cfg.RewardManagerState = solana.MustPublicKeyFromBase58(ProdRewardManagerState)
 		cfg.RewardManagerLookupTable = solana.MustPublicKeyFromBase58(ProdRewardManagerLookupTable)
